requirements.txt
matplotlib==3.7.5
natsort==8.4.0
numpy==1.24.4
onnx==1.19.1
onnxruntime_gpu==1.19.2
opencv_python==4.10.0.84
Orange==2.7.8
pandas==2.0.3
Pillow==12.0.0
scikit_learn==1.3.2
scipy==1.16.3
skimage==0.0
tensorboardX==2.6.2.2
tensorboardX==2.6.4
torch==2.4.1
torchvision==0.19.1
tqdm==4.65.2
